code moves
Moving the config and appname getters to the main class for reusability.

Sadly, the browser monitoring doesn't allow setting the app name in a way that's compatible, and there really isn't any use for the config method except in APM yet, but this is the right direction to go overall.

I also moved the Browser monitoring UI to the Settings tab, instead of Plugins.